Refactored FiniteFloat and FiniteDouble into Finites.scala.
